+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### To Release v0.6.0
 
+* [vc:fix] handle AM prefix as changes
+* [project] refresh project files viewer when select a new project
+
 ### v0.5.0
 
 * [vc:fix] c+k can trigger repeated commit editor